requirements.txt
requests
pyOpenSSL
requests_ntlm
dnspython
